Add file check

+++ BeatSaberMarkupLanguage/Plugin.cs
@@ -42,20 +42,23 @@ namespace BeatSaberMarkupLanguage
             SceneManager.activeSceneChanged += OnActiveSceneChanged;
             config = conf.Generated<Config>();
 
-            // Old Config Migration
 
+            // Old Config Migration
             Task.Run(() =>
             {
                 var folder = Path.Combine(UnityGame.UserDataPath, "BSML.ini");
-                string[] lines = File.ReadAllLines(folder);
-                string pinnnedModsLine = lines.FirstOrDefault(x => x.StartsWith("Pinned Mods"));
-                var splitLine = pinnnedModsLine.Split('=');
-                if (splitLine.Length > 1)
+                if (File.Exists(folder))
                 {
-                    var mods = splitLine[1].Split(',');
-                    config.PinnedMods.AddRange(mods.Where(x => !string.IsNullOrWhiteSpace(x) && x != " " && !config.PinnedMods.Contains(x)));
+                    string[] lines = File.ReadAllLines(folder);
+                    string pinnnedModsLine = lines.FirstOrDefault(x => x.StartsWith("Pinned Mods"));
+                    var splitLine = pinnnedModsLine.Split('=');
+                    if (splitLine.Length > 1)
+                    {
+                        var mods = splitLine[1].Split(',');
+                        config.PinnedMods.AddRange(mods.Where(x => !string.IsNullOrWhiteSpace(x) && x != " " && !config.PinnedMods.Contains(x)));
+                    }
+                    File.Delete(folder);
                 }
-                File.Delete(folder);
             });
         }
